# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Food Delivery Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalRevenue: number;
  totalOrders: number;
  totalCustomers: number;
  totalRestaurants: number;
  totalDrivers: number;
  activeOrders: number;
  avgOrderValue: number;
  revenueGrowth: number;
  ordersGrowth: number;
  customersGrowth: number;
}

export interface RevenueTrendPoint {
  date: string;
  revenue: number;
  orders: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface RestaurantRevenue {
  id: number;
  name: string;
  revenue: number;
  orders: number;
}

export interface HourlyOrders {
  hour: number;
  orders: number;
}

export interface Order {
  id: number;
  customerId: number;
  restaurantId: number;
  driverId?: number | null;
  status: string;
  totalAmount: number;
  deliveryAddress: string;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
  customerName: string;
  restaurantName: string;
  driverName?: string | null;
}

export interface CreateOrderRequest {
  customerId: number;
  restaurantId: number;
  driverId?: number | null;
  totalAmount: number;
  deliveryAddress: string;
  notes?: string;
}

export interface UpdateOrderStatusRequest {
  status: string;
  driverId?: number | null;
}

export interface OrderListResponse {
  data: Order[];
  total: number;
  page: number;
  limit: number;
}

export interface Customer {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  address?: string | null;
  totalOrders: number;
  totalSpent: number;
  createdAt: string;
}

export interface CreateCustomerRequest {
  name: string;
  email: string;
  phone?: string;
  address?: string;
}

export interface UpdateCustomerRequest {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
}

export interface CustomerListResponse {
  data: Customer[];
  total: number;
  page: number;
  limit: number;
}

export interface Restaurant {
  id: number;
  name: string;
  cuisine: string;
  address: string;
  phone?: string | null;
  email?: string | null;
  isActive: boolean;
  rating?: number | null;
  totalOrders: number;
  totalRevenue: number;
  createdAt: string;
}

export interface CreateRestaurantRequest {
  name: string;
  cuisine: string;
  address: string;
  phone?: string;
  email?: string;
  rating?: number;
}

export interface UpdateRestaurantRequest {
  name?: string;
  cuisine?: string;
  address?: string;
  phone?: string;
  email?: string;
  isActive?: boolean;
  rating?: number;
}

export interface RestaurantListResponse {
  data: Restaurant[];
  total: number;
  page: number;
  limit: number;
}

export interface Driver {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  vehicle?: string | null;
  licensePlate?: string | null;
  status: string;
  rating?: number | null;
  totalDeliveries: number;
  totalEarnings: number;
  createdAt: string;
}

export interface CreateDriverRequest {
  name: string;
  email: string;
  phone?: string;
  vehicle?: string;
  licensePlate?: string;
  rating?: number;
}

export interface UpdateDriverRequest {
  name?: string;
  email?: string;
  phone?: string;
  vehicle?: string;
  licensePlate?: string;
  status?: string;
  rating?: number;
}

export interface DriverListResponse {
  data: Driver[];
  total: number;
  page: number;
  limit: number;
}

export interface MenuItem {
  id: number;
  restaurantId: number;
  name: string;
  description?: string | null;
  price: number;
  category?: string | null;
  isAvailable: boolean;
  restaurantName: string;
}

export interface CreateMenuItemRequest {
  restaurantId: number;
  name: string;
  description?: string;
  price: number;
  category?: string;
}

export interface UpdateMenuItemRequest {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  isAvailable?: boolean;
}

export interface MenuItemListResponse {
  data: MenuItem[];
  total: number;
  page: number;
  limit: number;
}

export interface Payment {
  id: number;
  orderId: number;
  amount: number;
  method: string;
  status: string;
  transactionId?: string | null;
  createdAt: string;
  customerName: string;
  restaurantName: string;
}

export interface PaymentListResponse {
  data: Payment[];
  total: number;
  page: number;
  limit: number;
}

export type ListOrdersParams = {
  page?: number;
  limit?: number;
  status?: string;
  restaurantId?: number;
};

export type ListCustomersParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type ListRestaurantsParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type ListDriversParams = {
  page?: number;
  limit?: number;
  status?: string;
};

export type ListMenuItemsParams = {
  restaurantId?: number;
  page?: number;
  limit?: number;
};

export type ListPaymentsParams = {
  page?: number;
  limit?: number;
  status?: string;
};