# lib/api-zod/src/generated/types/updateMenuItemRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Food Delivery Admin Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateMenuItemRequest {
  name?: string;
  description?: string;
  price?: number;
  category?: string;
  isAvailable?: boolean;
}